// src/app/shared/models/commons/user-roles.ts
export enum Roles {
    
    ACCOUNTS = 'Accounts',
    ACCOUNTSADMIN = 'AccountsAdmin',
    ACCOUNTSSUPERADMIN = 'AccountsSuperAdmin',

    ADMINISTRATION = 'Administration',
    SUPERADMIN = 'SuperAdmin',
    SUPERADMINISTRATOR = 'SuperAdministrator',

    ACADEMICS = 'Academics',
    ACADEMICSADMIN = 'AcademicsAdmin',
    ACADEMICSSUPERADMIN = 'AcademicsSuperAdmin',

    DEVELOPERS = 'Developers',

    EXAMINATIONS = 'Examinations',
    EXAMINATIONSADMIN = 'ExaminationsAdmin',
    EXAMINATIONSSUPERADMIN = 'ExaminationsAdmin',

    FACULTY = 'Faculty',

    HR = 'HumanResources',
    HRAdmin = 'HumanResourcesAdmin',
    HRSuperAdmin = 'HumanResourcesSuperAdmin',
    HRReadWriteAccess = 'HumanResourcesReadWriteAccess',

    ITSUPPORTS = 'ITSupports',
    ITSUPPORTSADMIN = 'ITSupportsAdmin',
    ITSUPPORTSSUPERADMIN = 'ITSupportsSuperAdmin',

    LEADS = 'Leads',
    LEADSADMIN = 'LeadsAdmin',
    LEADSSUPERADMIN = 'LeadsSuperAdmin',
    
    MASTERS = 'Masters',
    MASTERSADMIN = 'MastersAdmin',
    MASTERSSUPERADMIN = 'MastersSuperAdmin',

    STUDENT = 'Student',
     
    VIRTUALLEARN = 'VirtualLearn',
    VIRTUALLEARNADMIN = 'VirtualLearnAdmin', 
    VIRTUALLEARNSUPERADMIN = 'VirtualLearnSuperAdmin',
}